Change package to build grub-redhat
This changes the upstream of the PKGBUILD, and adds various patches and
options for fixing issues during the build.

The result is an almost identical grub package, but compiles from
redhats sources instead of the upstream ones. This promises to be more
straigtforward when working with secure boot, as well as some additional
tweaks (like silent / flicker free boot)

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>