Add wrapper for installing a secureboot EFI binary
When booting with secureboot grub is not able to hotload modules /
drivers, so we have to include those in the main EFI binary, which gets
signed by the user.

Upstream does this by generating the binary during the packaging phase
and just rolling it out to the user that way. But that is not very
arch-like. Instead, we wrap grub-install with our own script
(grub-redhat-install), and inject the modules that should be included
dynamically.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>